+++ pytorch_NN/models/Alexnet.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-# @Time    : 2022/10/3 4:02
+# @Time    : 2022/10/3 4:59
 # @Author  : 石鑫磊
 # @Site    : 
-# @File    : __init__.py.py
+# @File    : Alexnet.py
 # @Software: PyCharm 
 # @Comment :
